feat: 削除ページを構成
