+++ src/app/components/layout/header/header.tsx
@@ -20,7 +20,7 @@ export default async function Header() {
         <div className="flex h-full w-[95%] items-center justify-between lg:w-[70%]">
           <Link href="/">
             <div className="flex items-center gap-2">
-              <Image src={"/image/logo.svg"} alt={"og_title"} width={40} height={40} />
+              <Image src={"/image/logo.svg"} alt={"og_title"} width={30} height={30} />
               <h1 className="text-[25px] font-bold">Tech Path</h1>
             </div>
           </Link>
@@ -37,7 +37,7 @@ export default async function Header() {
                 <PopoverContent align="end" className="w-[170px] p-0">
                   <ul>
                     <li className="border-b p-4 font-bold">{session.user.name}</li>
-                    <Link href={"/myPage"}>
+                    <Link href={"/my-page"}>
                       <li className="flex cursor-pointer items-center gap-3 border-b p-4">
                         <User size={15} />
                         自分の記事
