+++ src/app/components/myPage/delete-alert.tsx
@@ -14,9 +14,9 @@ const DeleteArert: React.FC<DeleteArertProps> = ({ isOpen, message, onClose, onC
   if (!isOpen) return null;
 
   return (
-    <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
+    <div className="fixed inset-0 z-10 flex items-center justify-center bg-black bg-opacity-50">
       <div className="rounded bg-white p-6 shadow-md">
-        <p className="mb-4 text-lg">{message}</p>
+        <p className="mb-4">{message}</p>
         <div className="flex justify-center gap-4">
           <Button onClick={onClose} className="rounded bg-gray-300 px-4 py-2 text-black hover:bg-gray-400">
             キャンセル
